Closes #2958 rename akstats to akscipy
