+++ pytest.ini
@@ -2,8 +2,8 @@
 filterwarnings =
     ignore:Version mismatch between client .*
 testpaths =
-    tests/akmath/akmath_test.py
-    tests/akstats/akstats_test.py
+    tests/akscipy/math_test.py
+    tests/akscipy/akscipy_test.py
     tests/alignment_tests.py
     tests/array_view_test.py
     tests/bitops_test.py
